+++ src/ball_tracking/BallTrackingKalmanFilterExtended2.h
@@ -54,7 +54,7 @@ namespace dayan
 		}
 
 		// Update Matrix A
-		void update_A(const int& dt) override
+		void update_A(const float& dt) override
 		{
 			auto dt2_2 = dt * dt / 2;
