+++ src/ball_tracking/BallTrackingKalmanFilterUnscented.h
@@ -17,7 +17,7 @@ namespace dayan
 	public:
 		// Constructor
 		BallTrackingKalmanFilterUnscented()
-			: KalmanFilterUnscented(6, 5)
+				: KalmanFilterUnscented(6, 5)
 		{
 			Q = 1e-6 * cv::Mat::eye(n, n, CV_32F);
 			R = 1 * cv::Mat::eye(m, m, CV_32F);
@@ -38,27 +38,27 @@ namespace dayan
 			auto Y = y * Z;
 
 			this->X = (cv::Mat_<float>(n, 1)
-				<<
-				X,
-				Y,
-				Z,
-				0,
-				0,
-				0
+					<<
+					X,
+					Y,
+					Z,
+					0,
+					0,
+					0
 			);
 		}
 
 		// Update Matrix A
-		void update_A(const int& dt) override
+		void update_A(const float& dt) override
 		{
 			A = (cv::Mat_<float>(n, n)
-				<<
-				1, 0, 0, dt, 0, 0,
-				0, 1, 0, 0, dt, 0,
-				0, 0, 1, 0, 0, dt,
-				0, 0, 0, 1, 0, 0,
-				0, 0, 0, 0, 1, 0,
-				0, 0, 0, 0, 0, 1
+					<<
+					1, 0, 0, dt, 0, 0,
+					0, 1, 0, 0, dt, 0,
+					0, 0, 1, 0, 0, dt,
+					0, 0, 0, 1, 0, 0,
+					0, 0, 0, 0, 1, 0,
+					0, 0, 0, 0, 0, 1
 			);
 		}
 
@@ -79,12 +79,12 @@ namespace dayan
 			auto Z2 = Z * Z;
 
 			h = (cv::Mat_<float>(5, 1)
-				<<
-				X / Z, // x
-				Y / Z, // y
-				Rm / Z, // r
-				-X * dZ / Z2 + dX / Z, // dx
-				-Y * dZ / Z2 + dY / Z // dy
+					<<
+					X / Z, // x
+					Y / Z, // y
+					Rm / Z, // r
+					-X * dZ / Z2 + dX / Z, // dx
+					-Y * dZ / Z2 + dY / Z // dy
 			);
 		}
 	};
